2002-09-14	Joel Sherrill <joel@OARcorp.com>
	* src/dosfs/fat.c, src/imfs/deviceio.c, src/imfs/imfs.h,
	src/imfs/imfs_directory.c, src/imfs/imfs_fchmod.c,
	src/imfs/memfile.c: Eliminated warnings.  A lot of this focus
	was on changing read and write entry points to return ssize_t
	as read(2) and write(2) now do.

+++ cpukit/libfs/src/imfs/deviceio.c
@@ -130,7 +130,7 @@ int device_close(
  *  This handler maps a read() operation onto rtems_io_read().
  */
 
-int device_read(
+ssize_t device_read(
   rtems_libio_t *iop,
   void          *buffer,
   unsigned32     count
@@ -158,7 +158,7 @@ int device_read(
   if ( status )
     return rtems_deviceio_errno(status);
 
-  return args.bytes_moved;
+  return (ssize_t) args.bytes_moved;
 }
 
 /*
@@ -167,7 +167,7 @@ int device_read(
  *  This handler maps a write() operation onto rtems_io_write().
  */
 
-int device_write(
+ssize_t device_write(
   rtems_libio_t *iop,
   const void    *buffer,
   unsigned32     count
@@ -195,7 +195,7 @@ int device_write(
   if ( status )
     return rtems_deviceio_errno(status);
 
-  return args.bytes_moved;
+  return (ssize_t) args.bytes_moved;
 }
 
 /*
